Adding RabbitMQ to make communication between Account API and Transaction API

+++ Src/Services/Transaction/Transaction.API/Models/Requests/WithDrawalRequest.cs
@@ -4,7 +4,7 @@
     {
         public string Trans_Author { get; set; }
         public string Trans_Description { get; set; }
-        public decimal Trans_Amount { get; set; }
+        public double Trans_Amount { get; set; }
         public Int64 Trans_Debited_Acc { get; set; }
     }
 }
